Navbar and Create Page

+++ backend/controllers/product.controller.js
@@ -72,6 +72,14 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({
+      success: false,
+      message: "Invalid Product ID",
+    });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({
@@ -80,9 +88,9 @@ export const deleteProduct = async (req, res) => {
     });
   } catch (error) {
     console.log("Error in Delete product:", error.message);
-    res.status(404).json({
+    res.status(500).json({
       success: false,
-      message: "Product not found",
+      message: "Server Error",
     });
   }
 };
